LiveSync: Move initial export utils to separate file
